feat: implement set count of users into group

+++ ChatZone.Api/Features/GroupMembers/LeaveGroup/LeaveGroupHandler.cs
@@ -1,5 +1,6 @@
 ﻿using ChatZone.Context;
 using ChatZone.Core.Extensions;
+using ChatZone.Core.Extensions.Exceptions;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -12,6 +13,10 @@ public class LeaveGroupHandler(
     public async Task<Result<IActionResult>> Handle(LeaveGroupRequest request, CancellationToken cancellationToken)
     {
         var groupMember = await dbContext.GroupMembers.SingleOrDefaultAsync(x => x.IdGroupMember == request.IdPerson, cancellationToken);
+        
+        var groupChat = await dbContext.GroupChats.SingleOrDefaultAsync(x => x.IdGroupChat == request.IdChat, cancellationToken);
+        if(groupChat is null) return Result<IActionResult>.Failure(new NotFoundException("Group chat is not found!"));
+        groupChat.UserCount -= 1;
         
         dbContext.GroupMembers.Remove(groupMember!);
         await dbContext.SaveChangesAsync(cancellationToken);
